+++ dune/stuff/common/parallel/threadmanager.hh
@@ -33,7 +33,7 @@ struct ThreadManager
   size_t thread();
 
   //! set maximal number of threads available during run
-  void set_max_threads(const int count);
+  void set_max_threads(const size_t count);
 
   ~ThreadManager() = default;
 
@@ -42,7 +42,7 @@ private:
   //! init tbb with given thread count, prepare Eigen for smp if possible
   ThreadManager();
 
-  int max_threads_;
+  size_t max_threads_;
 #if HAVE_TBB
   std::unique_ptr<tbb::task_scheduler_init> tbb_init_;
 #endif
